requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ── Templates ──────────────────────────────────────────────────────────────
jinja2==3.1.4

# ── Database — async SQLAlchemy + asyncpg for Neon PostgreSQL ──────────────
sqlalchemy==2.0.36
asyncpg==0.30.0

# ── Email — async SMTP ─────────────────────────────────────────────────────
aiosmtplib==3.0.2

# ── Config & validation ────────────────────────────────────────────────────
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# ── Utilities ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1
python-multipart==0.0.12     # required for FastAPI form data
